Move tests to separate directory
